split registering and handling of the slash commands
